requirements.txt
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.4
Django==5.2.1
django-admin-thumbnails==0.2.9
django-allauth==65.9.0
django-extensions==4.1
django-filter==25.1
django-jazzmin==3.0.1
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
idna==3.10
packaging==25.0
pillow==11.2.1
pycparser==2.22
PyJWT==2.9.0
python-decouple==3.8
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
timedelta==2020.12.3
tzdata==2025.2
urllib3==2.4.0
whitenoise==6.9.0
